feat: change styling
