+++ frontend/src/components/Common/styles/common_styles.scss
@@ -9,7 +9,7 @@
 }
 
 .header {
-	margin-bottom: 2rem;
+	margin-bottom: 1rem;
 	text-align: center;
 
 	.header-link {
