+++ frontend/src/pages/SearchPage.tsx
@@ -1,6 +1,8 @@
 import { Header } from "../components/Common/Common";
 import { FaUpload } from "react-icons/fa6";
 import CourseSearchForm from "../components/Search/SearchForm";
+import StatsShowcase from "../components/Search/Stats";
+
 
 function SearchPage() {
 	return <>
@@ -14,6 +16,7 @@ function SearchPage() {
 				button_text: "Upload!"
 			}}
 		/>
+		<StatsShowcase />
 		<CourseSearchForm />
 	</>;
 }
