+++ frontend/src/types/backend.ts
@@ -106,7 +106,6 @@ export interface IEndpointTypes {
 		request: null;
 		response: {
 			total_papers: number;
-			recent_uploads: number;
 			total_courses: number;
 		}
 	},
